pyproject.toml
[project]
name = "adult-income"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib==1.3.2",
    "pandas==2.1.3",
    "scikit-learn==1.3.2",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "dvc>=3.62.0",
]
